refactor: SourceType Path 변경
