+++ src/main/java/org/codeiteam3/findex/indexinfo/repository/IndexInfoRespository.java
@@ -1,6 +1,6 @@
 package org.codeiteam3.findex.indexinfo.repository;
 
-import org.codeiteam3.findex.SourceType;
+import org.codeiteam3.findex.enums.SourceType;
 import org.codeiteam3.findex.indexinfo.entity.IndexInfo;
 import org.springframework.data.jpa.repository.JpaRepository;
